host bn losses pool tensor_nn comparison

+++ src/hardware/fpga/nn/fpga_losses.cpp
@@ -45,11 +45,11 @@ void fpga_cent(Tensor *A, Tensor *B, Tensor *C){
   } else {
     cl_int err;
     cl::Event event;
- 
+
     OCL_CHECK(err, err = kernel_cent.setArg(0, (A->fpga_ptr)));
     OCL_CHECK(err, err = kernel_cent.setArg(1, (B->fpga_ptr)));
     OCL_CHECK(err, err = kernel_cent.setArg(2, (C->fpga_ptr)));
-    OCL_CHECK(err, err = kernel_cent.setArg(3, A->size));
+    OCL_CHECK(err, err = kernel_cent.setArg(3, (long int)A->size));
 
     OCL_CHECK(err, err = q.enqueueTask(kernel_cent, NULL, &event));
     //  event.wait();
